auth dir にloginを移動させた
