+++ backend/src/auth/auth.controller.ts
@@ -4,6 +4,7 @@ import { ApiOperation, ApiTags } from '@nestjs/swagger';
 import { AuthService } from './auth.service';
 import { accessToken } from './types/auth.types';
 import { signUpDto } from './dto/signUp.dto';
+import { loginDto } from './dto/login.dto';
 // TODO front直したら消す
 type User = {
   id: string;
@@ -29,6 +30,21 @@ export class AuthController {
   @ApiOperation({ summary: 'signUp nori' })
   async signUp(@Body() dto: signUpDto): Promise<User> {
     const userData = await this.authService.signUp(dto);
+    return {
+      id: userData.id,
+      createdAt: userData.createdAt,
+      updatedAt: userData.updatedAt,
+      email: userData.email,
+      nickname: userData.username,
+      hashedPassword: userData.hashedPassword,
+    };
+  }
+
+  @HttpCode(HttpStatus.OK)
+  @Post('login')
+  @ApiOperation({ summary: 'login nori' })
+  async login(@Body() dto: loginDto): Promise<User> {
+    const userData = await this.authService.login(dto);
     return {
       id: userData.id,
       createdAt: userData.createdAt,
