+++ backend/src/user/user.service.ts
@@ -1,31 +1,8 @@
-import { ForbiddenException, Injectable } from '@nestjs/common';
-import { User } from '@prisma/client';
-import { Prisma } from '@prisma/client';
+import { Injectable } from '@nestjs/common';
 
 import { PrismaService } from '../prisma/prisma.service';
 
-import { loginDto, signUpDto } from './dto/user.dto';
 @Injectable()
 export class UserService {
   constructor(private prisma: PrismaService) {}
-
-  async login(dto: loginDto): Promise<User> {
-    console.log(dto);
-    const user = await this.prisma.user.findUnique({
-      where: {
-        email: dto.email,
-      },
-    });
-    if (!user) {
-      console.log('emailが間違っている');
-      throw new ForbiddenException('Email incorrect');
-    }
-    if (user.hashedPassword != dto.hashedPassword) {
-      console.log(user.hashedPassword, dto.hashedPassword);
-      console.log('passwordが間違っている');
-      throw new ForbiddenException('Password incorrect');
-    }
-    console.log('OK');
-    return user;
-  }
 }
